--- inner-tool/src/main/java/util/code/BaseObjectCodeGenerator.java
@@ -1,5 +1,0 @@
-package util.code;
-
-public interface BaseObjectCodeGenerator {
-    Code createObjectCode(Object object, int level, String variableType, String variableName);
-}
