+++ inner-tool/src/main/java/util/code/ObjectCodeGenerator.java
@@ -3,32 +3,23 @@ package util.code;
 import common.GenCodeRequest;
 import common.Settings;
 
-import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.Date;
 import java.util.HashMap;
-import java.util.IdentityHashMap;
 import java.util.List;
 import java.util.Map;
 
 import static util.CompatibilityUtil.putToList;
-import static util.code.ObjectCodeHelper.escape;
 import static util.code.ObjectCodeHelper.firstLower;
 import static util.code.ObjectCodeHelper.getSimpleName;
-import static util.code.ObjectCodeHelper.isWrapperType;
 
-public class ObjectCodeGenerator implements BaseObjectCodeGenerator {
+public class ObjectCodeGenerator {
     private final Object rootObj;
     private final Settings settings;
 
     private final String variableName;
     private final String variableType;
-
-    private final UniqueNameGenerator uniqueNameGenerator = new UniqueNameGenerator();
-    private final Map<Object, ObjectCode> existingObjectCode = new IdentityHashMap<>();
 
     public ObjectCodeGenerator(Object rootObj, GenCodeRequest genCodeRequest) {
         this.rootObj = rootObj;
@@ -39,7 +30,9 @@ public class ObjectCodeGenerator implements BaseObjectCodeGenerator {
     }
 
     public String genCode() {
-        String root = createObjectCode(this.rootObj, 0, this.variableType, this.variableName).getCode();
+        ObjectCodeGeneratorCore objectCodeGeneratorCore = new ObjectCodeGeneratorCore(settings);
+        String root = objectCodeGeneratorCore.createObjectCode(this.rootObj, 0, this.variableType, this.variableName).getCode();
+        Map<Object, ObjectCode> existingObjectCode = objectCodeGeneratorCore.getExistingObjectCode();
 
         List<ObjectCode> outputCodes = new ArrayList<>(existingObjectCode.values());
         Collections.sort(outputCodes, new Comparator<ObjectCode>() {
@@ -118,60 +111,6 @@ public class ObjectCodeGenerator implements BaseObjectCodeGenerator {
         // add empty line before new block with assignments or after real assignments
         if (ret.length() != 0 && !objectCode.hasEmptyAssignment() || assignmentJustAdded) {
             ret.append("\n");
-        }
-    }
-
-    public Code createObjectCode(Object object, int level, String variableType, String variableName) {
-        if (object == null || level > this.settings.getMaxLevel()) {
-            return new Code("null");
-        } else if (isWrapperType(object.getClass())) {
-            if (object instanceof Integer) {
-                if (object.equals(Integer.MAX_VALUE)) {
-                    return new Code("Integer.MAX_VALUE");
-                } else if (object.equals(Integer.MIN_VALUE)) {
-                    return new Code("Integer.MIN_VALUE");
-                }
-                return new Code(object.toString());
-            } else if (object instanceof Float) {
-                return new Code(object + "f");
-            } else if (object instanceof Long) {
-                if (object.equals(Long.MAX_VALUE)) {
-                    return new Code("Long.MAX_VALUE");
-                } else if (object.equals(Long.MIN_VALUE)) {
-                    return new Code("Long.MIN_VALUE");
-                }
-                return new Code(object + "L");
-            } else if (object instanceof Character) {
-                return new Code("'" + object + "'");
-            } else {
-                return new Code(object.toString());
-            }
-        } else if (object instanceof String) {
-            return new Code("\"" + escape((String) object) + "\"");
-        } else if (object instanceof Enum) {
-            return new Code(object.getClass().getSimpleName() + "." + object);
-        } else if (object instanceof Date) {
-            return new Code("new " + object.getClass().getSimpleName() + "(" + ((Date) object).getTime() + "L)");
-        } else if (object instanceof BigDecimal) {
-            return new Code("new " + object.getClass().getSimpleName() + "(" + object + ")");
-        } else if (object instanceof BigInteger) {
-            return new Code("new " + object.getClass().getSimpleName() + "(\"" + object + "\")");
-        } else {
-            ObjectCode existed = existingObjectCode.get(object);
-            if (existed != null) {
-                if (existed.constructorLevel < level) {
-                    existed.constructorLevel = level;
-                }
-                return new Code(existed.referenceName, existed);
-            } else {
-                String referenceName = variableName != null
-                        ? uniqueNameGenerator.createUniqueName(variableName)
-                        : uniqueNameGenerator.genReferenceName(object.getClass());
-                ObjectCode objectCode = new ObjectCode(this, settings, level, referenceName, object, variableType);
-                existingObjectCode.put(object, objectCode);
-                objectCode.walkObjectsTree();
-                return new Code(referenceName, objectCode);
-            }
         }
     }
 }
